Remove annoying CoffeeScript asset
Now we should be able to test the engine in the `test/dummy` directory
by running `rails s` as usual.

This asset was added by Action Cable. We don't want to make all
stakeholders of Alexandrie install the asset handler for CoffeeScript,
since we rarely use it anyway.

Such is the life of the bleeding edge ¯\_(ツ)_/¯

Signed-off-by: Emil Henry <emil@plaimi.net>